requirements.txt
atlassian-python-api==1.11.19
pandas==0.24.0
lxml==4.3.0
